// apps/home.app.weather/src/services/geolocation/get.geojs.io.types.ts
export interface GeoJsonResponse {
  timezone: string;
  city: string;
  asn: number;
  country: string;
  area_code: string;
  organization_name: string;
  country_code: string;
  country_code3: string;
  continent_code: string;
  organization: string;
  region: string;
  ip: string;
  longitude: string;
  accuracy: number;
  latitude: string;
}